SBERDOMA-348 moved `formatDate` to `ticket` domain
