+++ apps/condo/domains/ticket/components/TicketChanges/TicketChange.tsx
@@ -3,7 +3,7 @@ import { Row, Col, Typography, Tooltip } from 'antd'
 import { has } from 'lodash'
 import styled from '@emotion/styled'
 import { TicketChange as TicketChangeType } from '@app/condo/schema.d'
-import { formatDate } from '../../../common/utils/helpers'
+import { formatDate } from '../../utils/helpers'
 import { useIntl } from '@core/next/intl'
 import { PhoneLink } from '@condo/domains/common/components/PhoneLink'
 import { green } from '@ant-design/colors'
